chore(js): update memoized functions 🤖
